+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.4.0]
 #### Added
  * Allow the parameters user, application name and comment to queries
  * Add tests (test_pragmas and test_pragmas test_pragmas_not_comment_free) in the query.py file.
+ 
+#### Fixed
+ * Problems when API CAll query has not "to", now added be default "now()" if not to in no stream mode
 
 ## [1.3.1] - 2018-12-04
 #### Fixed
@@ -21,7 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  
 #### Fixed
  * ConnectionError in python 2
- * tests of Sender
+ * tests of Sender  
  * Close order of API/Buffer
  
 #### Changed